+++ lib/zaikio/mission_control/specialist.rb
@@ -1,0 +1,9 @@
+module Zaikio
+  module MissionControl
+    class Specialist < Base
+      # Associations
+      has_many :availabilities, class_name: "Zaikio::MissionControl::Availability",
+                                uri: "specialists/:machine_id/availabilities"
+    end
+  end
+end