+++ test/zaikio/mission_control_test.rb
@@ -192,4 +192,33 @@ class Zaikio::MissionControlTest < ActiveSupport::TestCase
       end
     end
   end
+
+  test "create single availability" do
+    WebMock.disable_net_connect!
+    machine_id = "f44b5c9c-2c58-4f31-a02f-632cce70320b"
+
+    stub_request(:post, "https://mc.zaikio.test/api/v1/machines/#{machine_id}/availabilities")
+      .with(body: { availability: {
+        mode: "available",
+        starts_at: 34_200,
+        ends_at: 63_000,
+        weekday: "monday",
+        capacity: 1
+      } }.to_json)
+      .to_return(status: 201, body: {
+        availability: { "id" => "e0815217-be2f-436e-915a-67bec7faf010" }
+      }.to_json, headers: { "Content-Type" => "application/json" })
+
+    Zaikio::MissionControl.with_token(token) do
+      availability = Zaikio::MissionControl::Machine.new(id: machine_id).availabilities.create(
+        mode: "available",
+        starts_at: 34_200,
+        ends_at: 63_000,
+        weekday: "monday",
+        capacity: 1
+      )
+
+      assert_equal "e0815217-be2f-436e-915a-67bec7faf010", availability.id
+    end
+  end
 end