+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add availabilities for machines and specifications
+
 ## [0.2.3]
 
 - Add basic api resource retrieval for `Jobs`
+
 ## [0.2.2]
 
 - Improve german translations